fix: resolve critical build breaks and architectural issues
- Remove kotlin.exclude that hid ALL background code from compilation
- Fix BackgroundSecureTool to extend SecureTool with correct ToolResult type
- Fix KoogRoutine to register phase tools and transition tools (was missing them)
- Unify toKoogTool() — single implementation with rich JSON Schema types
- Add ToolResult.Structured handling to all exhaustive when expressions
- Install Koog Persistence feature in PhaseAwareAgent for agent checkpoints
- Persist serializedState and toolCallCounts in RoomSessionStore
- Fix WorkManagerObserver scope leak — cancels scope on terminal state
- Fix PhaseSession.isRunning dead code (also/let pattern never collected)
- Replace System.currentTimeMillis() with KMP-safe Clock.System
- Delete empty koog-compose-android-workmanager module (code lives in device)
- Remove 7 dead stub files with only comments
