+++ .qwen/settings.json.orig
@@ -1,0 +1,7 @@
+{
+  "permissions": {
+    "allow": [
+      "WebFetch(docs.koog.ai)"
+    ]
+  }
+}
