--- koog-compose-android-workmanager/build.gradle.kts
@@ -1,42 +1,0 @@
-
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-
-plugins {
-    alias(libs.plugins.kotlinMultiplatform)
-    alias(libs.plugins.androidLibrary)
-    alias(libs.plugins.dokka)
-    id("publish-convention")
-}
-
-kotlin {
-    explicitApi()
-
-    androidTarget {
-        compilerOptions {
-            jvmTarget.set(JvmTarget.JVM_11)
-        }
-    }
-
-    sourceSets {
-        commonMain.dependencies {
-            implementation(project(":koog-compose-core"))
-            implementation(libs.kotlinx.coroutines.core)
-            implementation(libs.kotlinx.serialization.json)
-        }
-        androidMain.dependencies {
-            implementation(libs.androidx.work.runtime.ktx)
-        }
-    }
-}
-
-android {
-    namespace = "io.github.koogcompose.android.workmanager"
-    compileSdk = 35
-    defaultConfig {
-        minSdk = 24
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-}
