--- koog-compose-android-workmanager/gradle.properties
@@ -1,4 +1,0 @@
-# Per-module POM metadata (overrides root where needed)
-POM_ARTIFACT_ID=koog-compose-android-workmanager
-POM_NAME=Koog Compose Android WorkManager
-POM_DESCRIPTION=Android WorkManager integration for background task scheduling in koog-compose applications.
