+++ koog-compose-core/src/commonMain/kotlin/io/github/koogcompose/session/PhaseSession.kt
@@ -52,22 +52,18 @@ public class PhaseSession<S>(
     private val _activity = MutableStateFlow<AgentActivity>(AgentActivity.Idle)
     override val activity: StateFlow<AgentActivity> = _activity.asStateFlow()
 
-    private val _isRunningDerived = MutableStateFlow(false)
-
+    private val _isRunning = MutableStateFlow(false)
 
     private val _activityDetail = MutableStateFlow("")
     override val activityDetail: StateFlow<String> = _activityDetail.asStateFlow()
 
     // Derived from activity — backward-compatible with existing code that reads isRunning.
-    override val isRunning: StateFlow<Boolean> = MutableStateFlow(false).also { derived ->
+    override val isRunning: StateFlow<Boolean> = _isRunning.asStateFlow()
+
+    init {
         scope.launch {
-            _activity.collect { derived.value = it.isRunning }
+            _activity.collect { _isRunning.value = it.isRunning }
         }
-    }.let {
-        // Return the underlying StateFlow via the activity map.
-        // We use a dedicated MutableStateFlow updated via collect to avoid
-        // needing stateIn (which requires a lifecycle scope we don't own).
-        _isRunningDerived
     }
 
 
@@ -103,13 +99,6 @@ public class PhaseSession<S>(
     // ── Agent lifecycle ────────────────────────────────────────────────────
 
     private var agent: AIAgent<String, String>? = null
-
-    init {
-        // Keep _isRunningDerived in sync with _activity.
-        scope.launch {
-            _activity.collect { _isRunningDerived.value = it.isRunning }
-        }
-    }
 
     // ── Public API ─────────────────────────────────────────────────────────
