--- koog-compose-device/src/androidMain/kotlin/io/github/koogcompose/device/background/BackgroundJobProvider.kt
@@ -1,1 +1,0 @@
-// DEPRECATED: Moved to :koog-compose-core io.github.koogcompose.background.BackgroundJobProvider
