+++ koog-compose-device/src/androidMain/kotlin/io/github/koogcompose/device/background/BackgroundSecureTool.kt
@@ -3,10 +3,18 @@ package io.github.koogcompose.device.background
 import android.content.Context
 import io.github.koogcompose.background.BackgroundJobProvider
 import io.github.koogcompose.event.KoogEventBus
+import io.github.koogcompose.tool.SecureTool
+import io.github.koogcompose.tool.ToolResult
+import io.github.koogcompose.tool.PermissionLevel
+import kotlinx.serialization.json.JsonObject
 import java.util.UUID
 
 /**
  * Abstract base for AI tools that delegate heavy work to a background job.
+ *
+ * Implement this class to create tools that run via WorkManager (Android) or
+ * background sessions (iOS). The LLM sees this as a normal tool, but execution
+ * is offloaded to a background worker.
  *
  * @param context     Application context — used for permission checks and WorkManager.
  * @param jobProvider The [BackgroundJobProvider] to enqueue work through.
@@ -16,14 +24,15 @@ public abstract class BackgroundSecureTool(
     private val context: Context,
     private val jobProvider: BackgroundJobProvider,
     private val eventBus: KoogEventBus,
-) {
-    /** Unique tool name exposed to the LLM. */
-    public abstract val toolName: String
+) : SecureTool {
 
-    /**
-     * Android permission strings that must be granted before the job is enqueued.
-     */
+    /** Unique tool name exposed to the LLM ( SecureTool.name ). */
+    public abstract override val name: String
+
+    /** Android permission strings that must be granted before the job is enqueued. */
     public open val requiredPermissions: List<String> = emptyList()
+
+    override val permissionLevel: PermissionLevel = PermissionLevel.SAFE
 
     /**
      * Builds the [Map] of key-value pairs passed into the background worker.
@@ -31,22 +40,23 @@ public abstract class BackgroundSecureTool(
     public abstract fun buildInputData(args: Map<String, String>): Map<String, String>
 
     /**
-     * Entry point called by the tool dispatch layer.
+     * Entry point called by the tool dispatch layer. Enqueues a background job
+     * and returns immediately. The LLM sees a success message with the job ID.
      */
-    public fun execute(args: Map<String, String>): ToolResult {
+    final override suspend fun execute(args: JsonObject): ToolResult {
         val missing = requiredPermissions.filter { permission ->
             context.checkSelfPermission(permission) !=
                 android.content.pm.PackageManager.PERMISSION_GRANTED
         }
 
         if (missing.isNotEmpty()) {
-            return ToolResult.Error(
-                "Cannot run $toolName: missing permissions: ${missing.joinToString()}"
+            return ToolResult.Failure(
+                "Cannot run $name: missing permissions: ${missing.joinToString()}"
             )
         }
 
-        val jobId = "$toolName-${UUID.randomUUID()}"
-        val inputData = buildInputData(args)
+        val jobId = "$name-${UUID.randomUUID()}"
+        val inputData = buildInputData(args.mapValues { (_, v) -> v.toString() })
 
         jobProvider.enqueue(jobId, inputData)
 
@@ -54,9 +64,4 @@ public abstract class BackgroundSecureTool(
 
         return ToolResult.Success("Job [$jobId] started in background.")
     }
-}
-
-public sealed class ToolResult {
-    public data class Success(val message: String) : ToolResult()
-    public data class Error(val message: String) : ToolResult()
 }
