--- koog-compose-device/src/androidMain/kotlin/io/github/koogcompose/device/session/KoogDatabase.android.kt
@@ -1,1 +1,0 @@
-// REDUNDANT - Consolidated into :koog-compose-session-room
