--- koog-compose-device/src/androidMain/kotlin/io/github/koogcompose/device/session/KoogDatabase.kt
@@ -1,1 +1,0 @@
-// Moved to :koog-compose-session-room
