--- koog-compose-device/src/commonMain/kotlin/io/github/koogcompose/device/session/KoogDatabase.kt
@@ -1,1 +1,0 @@
-// DELETED - Consolidated into :koog-compose-session-room
