--- koog-compose-device/src/commonMain/kotlin/io/github/koogcompose/device/session/RoomSessionStore.kt
@@ -1,1 +1,0 @@
-// Moved to :koog-compose-session-room
