--- koog-compose-device/src/commonMain/kotlin/io/github/koogcompose/device/session/Roomentities.kt
@@ -1,1 +1,0 @@
-// Moved to :koog-compose-session-room
