--- koog-compose-device/src/iosMain/kotlin/io/github/koogcompose/device/session/KoogDatabase.ios.kt
@@ -1,1 +1,0 @@
-// Moved to :koog-compose-session-room
